+++ utopia-remix/app/store.tsx
@@ -42,7 +42,7 @@ interface ProjectsStoreActions {
   setGridView: (gridView: boolean) => void
   addOperation: (operation: Operation, key: string) => void
   removeOperation: (key: string) => void
-  updateOperation: (key: string, errored: boolean) => void
+  updateOperation: (key: string, data: { errored: boolean }) => void
 }
 
 type ProjectsStore = ProjectsStoreState & ProjectsStoreActions
@@ -84,11 +84,14 @@ export const useProjectsStore = create<ProjectsStore>()(
             operations: operations.filter((other) => other.key !== key),
           }))
         },
-        updateOperation: (key, errored) => {
+        updateOperation: (key, data) => {
           return set(({ operations }) => ({
             operations: operations.map((other) => {
               if (other.key === key) {
-                return { ...other, errored: errored }
+                return {
+                  ...other,
+                  errored: data.errored,
+                }
               }
               return other
             }),
